Finished file upload capability in Angular example
